requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
hf-xet==1.1.3
    # via huggingface-hub
huggingface-hub==0.32.4
    # via mlx-whisper
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
llvmlite==0.44.0
    # via numba
markupsafe==3.0.2
    # via jinja2
mlx==0.26.1
    # via mlx-whisper
mlx-whisper==0.4.2
    # via -r requirements.in
more-itertools==10.7.0
    # via mlx-whisper
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numba==0.61.2
    # via
    #   -r requirements.in
    #   mlx-whisper
numpy==2.2.6
    # via
    #   -r requirements.in
    #   mlx-whisper
    #   numba
    #   scipy
packaging==25.0
    # via huggingface-hub
pyaudio==0.2.14
    # via -r requirements.in
pyobjc-core==11.0
    # via
    #   -r requirements.in
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-cocoa
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
pyobjc-framework-applicationservices==11.0
    # via -r requirements.in
pyobjc-framework-cocoa==11.0
    # via
    #   -r requirements.in
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
    #   rumps
pyobjc-framework-coretext==11.0
    # via pyobjc-framework-applicationservices
pyobjc-framework-quartz==11.0
    # via
    #   -r requirements.in
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
python-dotenv==1.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via huggingface-hub
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   tiktoken
rumps==0.4.0
    # via -r requirements.in
scipy==1.15.3
    # via mlx-whisper
sympy==1.14.0
    # via torch
tiktoken==0.9.0
    # via mlx-whisper
torch==2.7.1
    # via mlx-whisper
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   mlx-whisper
typing-extensions==4.14.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.4.0
    # via requests
